fix: Fix Chinese/English mode switching bug
When the input method is changed to English mode with 'shift' and
switch to the next input method with 'ctrl-shift', it requires pressing
the 'shift' button 2 times to change to the English mode.
The error is caused by forgetting to set the 'key_press_alt' back to
the correct value.