fix: Map GTK2 GtkStatusIcon code to GTK3
GTK3 does not support GtkStatusIcon related code.
The document suggestion is
"GtkStatusIcon has been deprecated in 3.14.
You should consider using notifications or more modern platform-specific APIs instead.
GLib provides the GNotification API which works well with GtkApplication
on multiple platforms and environments."
Ref:
https://docs.gtk.org/gtk3/class.StatusIcon.html

However, the notification does not fit the tray icon requirement.
And it needs a lot of code work to implement the same function
in a different way.
To make GTK3 compile successfully, the mappings are added.
It is a workaround and should be modified in the future.