fix: Fix errors of specific web inputs on Firefox
There are 2 different input errors:
1. The first Chinese character would be committed automatically.
For example: Facebook popup window for posting and chat box of Facebook
and LinkedIn.
2. The first Chinese preedit sequence can not be committed successfully.
For example: Telegram

These errors are fixed with reordering the signals of GTK IMContext.
However, the errors would not appear in other web input situations.

The relation of errors and reordering is not cleared yet.
But the reordering is reasonable since the preedit change should come
after the preedit is commited.
preedit-changed signal: "It is also emitted at the end of a preedit sequence"
Ref:
https://docs.gtk.org/gtk3/signal.IMContext.preedit-changed.html