CI: Extract code formatting as single step and use static clang-format
The code formatting of CI is extracted from building the Arch Linux
step for clarity.
The CI of Arch Linux would install clang-format version 15,
which would be slightly different with version 14
when handling the style of the comments.
The developers of the project should use the same version of
clang-format.