chore: Remove USE_TSIN macros and related configuration
The TSIN related code are scattered between the files of different
functions.
Some of them are controlled with USE_TSIN macro and others are
controlled with if-else statement detecting the current input method,
or both.

To avoid dirty source code, TSIN related code should be extracted and
placed in TSIN related files.
Besides, the code should be controlled with detecting the current input
method instead of USE_TSIN macros since it is not necessary.
Other input method does not use the macro mechanism like TSIN.
If the users do not want to use TSIN, they can simply disable it in the
setting page instead of disable it in the configure step, which most
users would not even know how to do it.

The commit would be the first step to isolate the TSIN related code.