refactor: Extract functions for setting widget color
Extract common functions for setting the foreground and backgound color
of the widgets.
The code should not be placed in win-gtab.c since other input method
window may use the functions as well.
The functions are renamed from "change" to "apply" since the color might
not be changed to another color based on the condition.

The function can be improved since it parses the color each time.