feat: Remove using environment variable HIME_DESTROY_WINDOW
Purpose:
The HIME_DESTROY_WINDOW environment variable would determine whether to
destroy the window or not when hiding it.
The window is the topmost widget for input methods showing the inputs.
There is no document or comment describing the environment variable
usage and purpose.
It is kind of a legacy code and should be removed for simplicity.

Modification:
Remove environment variable detection of HIME_DESTROY_WINDOW and the
related variables and conditional code.
The topmost window would be hidden instead of destroyed when calling the
hiding function.