fix: Fix TSIN window disappear issue
The TSIN input window has two parts.
One is the tsin part of preedit buffer.
The other one is the phone part.
When using backspace to make the phone part empty, it would hide the
TSIN input window no matter the preedit buffer of tsin part is empty or not.

Modification:
Add the condition checking of tsin part.
The condition would check whether the input of tsin part is empty or
not.
If it is empty, then it is acceptable to hide the TSIN input window.
Otherwise, it should be shown.