feat: Remove displaying tsin cursor at the end of preedit buffer
When first switching from Chinese to English in TSIN input method,
a cursor would be shown at the end of the preedit buffer, even if
switching back.
If there is no switching from Chinese to English, there would not have
an cursor at the end of preedit buffer in Chinese mode.
Therefore, remove the related code would make sure the consistency of
cursor behavior.