refactor: Rename function clrin_pho
The new function name reset_pho_structure() is better since 'clrin' is
not a good prefix and 'reset_pho_strucutre' means resetting the structure of
the PHO input method.

TODO:
The function contains the code for hiding the PHO input window.
The hiding should be extracted to another function in the future.