fix: Assign the correct method type to GTAB input method
In load_gtab_list(), it does not take care of the GTAB input method
correctly.
The GTAB input method should be assigned with the method type
method_type_GTAB.
The original code does not cause any problem since the detection of
method type is mostly implemented with switch-case and there is no
detection of method_type_GTAB but using default instead.