CI: Remove CentOS 7 CI testing
The latest GCC compiler version in CentOS 7 is 4.8.5.
In HIME, it uses gnu17 and gnu++17 as its language standard.
However, -std=gnu17 compiler option is not supported in such an obsolete
compiler.
Besides, the end-of-life of CentOS 7 is at 2024/6/30.
It would be easier to maintain the CI system if there is no special
request for testing on CentOS 7.